feat: add custom CSS support
This commit adds support for custom CSS in the preference service. It introduces a new property, `customCss`, in the `IPreferenceStore` interface and sets its initial value to an empty string. Additionally, it adds logic in the `WindowProcessManagementService` to insert the custom CSS when a window finishes loading and remove it when necessary.

Continues #571